refactor(tools/bash): move NON_INTERACTIVE_ENV and _truncate_output out of Tool base
These are bash/container-specific; polluting the abstract Tool base was
premature generalization. Both now live as module-level items in bash.py.
